requirements.txt
# Django & API
Django==5.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0

# Database
mysqlclient==2.2.0

# Celery & Redis
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# ML & AI
scikit-learn==1.3.2
xgboost==1.7.6
transformers==4.30.0
torch==2.0.1
pandas==2.0.3
numpy==1.24.3
joblib==1.3.2
pulp==2.7.0

# Web Scraping
facebook-scraper==0.2.59
beautifulsoup4==4.12.2
requests==2.31.0

# Nepali NLP (optional - install manually if needed)
# nepali-stemmer==1.0.0

# Visualization & Frontend
matplotlib==3.7.2
seaborn==0.12.2

# Utilities
python-decouple==3.8
Pillow==10.0.1
python-dateutil==2.8.2
pytz==2023.3

# Development
jupyter==1.0.0
notebook==6.5.4

# Deployment
gunicorn==21.2.0
whitenoise==6.5.0
dj-database-url==2.1.0